Massive changes around resizing of blocks

+++ admin/section-template-reordering.php
@@ -8,7 +8,8 @@
  */
 
 ?>
-<?php if ( Multiple_Content_Sections::$template_data[ $selected_template ]['blocks'] > 1 ) : ?>
+
+<?php if ( (int) Multiple_Content_Sections::$template_data[ $selected_template ]['blocks'] > 1 ) : ?>
 <div id="mcs-description" class="description notice notice-warning is-dismissible below-h2">
 	<p>
 		<?php esc_html_e( 'Reorder your content blocks by dragging and dropping.', 'linchpin-mcs' ); ?>
